Boarding the plane
There are n rows in the plane and three seats on the left and right in each row. The extreme seats (A and F) are at the window, the central ones (C and D) are at the aisle. Groups of one, two or three passengers come to check-in. They want to sit side by side, that is, on the same row and on the same side: left or right. For example, a group of two passengers can sit on seats B and C, but cannot sit on seats C and D because they are separated by an aisle, and also cannot sit on seats A and C, because then they will not be next to each other. In addition, one of the passengers in each group is very demanding – he wants to sit either at the window or at the aisle. Of course, each group of passengers wants to take seats in a row with as small a number as possible, because then they will get off the plane more quickly after landing. For each group of passengers, determine whether there are seats on the plane suitable for them.
The first line contains the number n (1≤n≤100) – the number of rows in the plane. Next, in n lines, the initial seating in the plane is entered in rows (from the first to the nth), where characters. (dot) indicates empty seats, the symbols # (grid) indicate occupied seats, and the symbols _ (underscore) indicate the passage between seats C and D of each row.
The next line contains the number m (1≤m≤100) - the number of passenger groups. Next in m
 the lines contain descriptions of passenger groups. The description format is as follows: num side position, where num is the number of passengers (number 1, 2 or 3), side is the desired side of the aircraft (left or right line), position is the desired seat of a demanding passenger (aisle or window line).
If the group can take seats that meet its requirements, then output the string Passengers can take seats: and a list of their seats in the format r o w l e t t e r, ordered in ascending order of the letter of the seat. Then print in n lines the resulting seating on the plane, in the format described above, and the seats occupied by the current group of passengers should be indicated by the symbol X. If the group cannot find seats that meet its requirements, then output the string Cannot fulfill passengers requirements.