Dynamic Programming Solution
Print the route of the maximum cost. 
There is a turtle in the upper left corner of a rectangular table measuring N×M. A number is written in each cell of the table. The turtle can move to the right or down, with the turtle's route ending in the lower right corner of the table.
Let's calculate the sum of the numbers recorded in the cells through which the turtle crawled (including the initial and final cells). Find the largest possible value of this amount and the route on which this amount is reached.