The Speleologist's Way
The cave is represented by a cube divided into N parts for each dimension.
Each cell can be either empty or completely filled with a stone.
Based on the position of the caver in the cave, it is required to find out what is the minimum number of
movements through the cells he needs to get to the surface.You can move from cell to cell only if they are 
both freeand have a common face.
The first line contains the number N (1 ≤ N ≤ 30). This is followed by N blocks.
The block consists of an empty line and N lines of N characters : #  - denotes a cell filled with stones, dot - a free cell.
The initial position of the caver is indicated by the capital letter S.
The first block represents the upper level of the cave, reaching any of its free cells means reaching the surface.
Access to the surface is always possible.
Print one number - the length of the path to the surface.